+++ lib/core/utils/app_router.dart
@@ -5,7 +5,7 @@ import 'package:doc_doc_app/features/home/data/repo/home_repo.dart';
 import 'package:doc_doc_app/features/home/presentation/manager/doctor_cubit/doctor_cubit.dart';
 import 'package:doc_doc_app/features/home/presentation/views/home_view.dart';
 import 'package:doc_doc_app/features/home/presentation/views/notification_view.dart';
-import 'package:doc_doc_app/features/recommendation/presentation/views/recommendation_view.dart';
+import 'package:doc_doc_app/features/home/presentation/views/recommendation_view.dart';
 import 'package:doc_doc_app/features/home/presentation/views/speciality_view.dart';
 import 'package:doc_doc_app/features/login/presentation/views/login_view.dart';
 import 'package:doc_doc_app/features/onboarding/presentation/views/onboarding_view.dart';
